cam_join + cam_info

+++ silo-server/src/main/java/pt/tecnico/sauron/silo/domain/Camera.java
@@ -2,20 +2,20 @@ package pt.tecnico.sauron.silo.domain;
 
 public class Camera {
     public String name;
-    public int latitude;
-    public int longitude;
+    public float latitude;
+    public float longitude;
 
-    public Camera(String name, int latitude, int longitude){
+    public Camera(String name, float latitude, float longitude){
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
@@ -27,11 +27,11 @@ public class Camera {
         this.name = name;
     }
 
-    public int getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
